fix: 修复 CI 编译错误——sendLog 主线程读书架数、durChapterIndex 用 Int32

+++ Features/Reader/VideoReadView.swift
@@ -178,7 +178,7 @@ struct VideoReadView: View {
     }
 
     private func saveRecord(index: Int) {
-        book.durChapterIndex = Int64(index)
+        book.durChapterIndex = Int32(index)
         try? book.managedObjectContext?.save()
     }
 }
